xassert: don't use XASSERT_EQ for XASSERT_{NOT_,}NULL
Currently, XASSERT_NOT_NULL breaks under pedantic mode when applied to
function pointers because of this warning:

ISO C forbids conversion of function pointer to object pointer type

This is because function pointers and pointers to anything else ("object
pointers") are not treated the same, and casting a function pointer to
void * is not allowed under the standard. You can, however, compare the
function pointer to NULL.

Although this is all very sad, use XASSERT instead of XASSERT_EQ for
XASSERT_NOT_NULL and XASSERT_NULL. Though we won't get to see the
pointer value, at least it will always work.