Revert "xassert: use kill instead of abort"
This erroneously kills parent processes when a process using xlib is run
as a child. This is undesireable, as it results in the a unit test
runner crashing instead of just the unit test! To handle the case of
killing zombie children (the reason this commit was put in place),
instead the unit test should kill children directly.